Use the same icon for all charge levels if we are in the charging status
From the designers (see bug comments):
"I think the shading inside the battery when charging is obscuring the
lightning bolt, and also giving the icon a visual style that is obviously
different from the text and discourages "reading" it together with the text.
So, I suggest leaving the brackets in place, but removing the shading from
the icon.

Fixes https://bugs.launchpad.net/indicator-power/+bug/824629