Don't prioritize discharging items with no time estimate that have more than 10% power remaining.
Devices with no time estimates are most likely low-power devices that
have long-lasting batteries, such as a mouse with AA batteries. For those
type of devices that contain batteries that last weeks, there is no value
in displaying their status in preference to devices that have a rapid
charge/discharge cycle.

However, there is value in knowing if the device has a battery that needs
replacing imminently, so only display it if it falls to a 10% charge or
under.